+++ eveuniverse/tests/test_models_4.py
@@ -1,5 +1,6 @@
 """Eve Entity tests."""
 
+from typing import Dict
 from unittest.mock import patch
 
 from eveuniverse.models import EveEntity
@@ -553,15 +554,37 @@ class TestEveEntityManagerFetchEntitiesByName(NoSocketsTestCase):
     def test_should_fetch_unknown_entities_from_esi_only(self, mock_esi):
         # given
         mock_esi.client.Universe.post_universe_ids.return_value = BravadoOperationStub(
-            []
+            {
+                "characters": [
+                    {"id": 9991, "name": "alpha"},
+                    {"id": 9992, "name": "bravo"},
+                ],
+                "corporations": [
+                    {"id": 9993, "name": "charlie"},
+                ],
+            }
         )
         create_eve_entity(
             id=1001, name="Bruce Wayne", category=EveEntity.CATEGORY_CHARACTER
         )
         # when
-        EveEntity.objects.fetch_by_names_esi(["Bruce Wayne", "unknown"])
+        result_qs = EveEntity.objects.fetch_by_names_esi(
+            ["Bruce Wayne", "alpha", "bravo", "charlie"]
+        )
         # then
         self.assertTrue(mock_esi.client.Universe.post_universe_ids.called)
+        _, kwargs = mock_esi.client.Universe.post_universe_ids.call_args
+        self.assertSetEqual(set(kwargs["names"]), {"alpha", "bravo", "charlie"})
+        objs: Dict[int, EveEntity] = {obj.id: obj for obj in result_qs}
+        self.assertSetEqual(set(objs.keys()), {1001, 9991, 9992, 9993})
+        self.assertEqual(objs[1001].name, "Bruce Wayne")
+        self.assertTrue(objs[1001].is_character)
+        self.assertEqual(objs[9991].name, "alpha")
+        self.assertTrue(objs[9992].is_character)
+        self.assertEqual(objs[9992].name, "bravo")
+        self.assertTrue(objs[9992].is_character)
+        self.assertEqual(objs[9993].name, "charlie")
+        self.assertTrue(objs[9993].is_corporation)
 
 
 class TestEveEntityProfileUrl(NoSocketsTestCase):